mask8x8::count_true can skip some steps
I originally stubbed this as a call to the Rust library's own popcount
function but had intended that to just be a placeholder while working on
something else.

The core-provided count_ones method is correct but does unnecessary work
because we already know there's at most one bit set in each octet. Instead
then, we can effectively start the typical tree-based popcount process
partway through and perform only the remaining three steps to reach a
popcount for the whole u64 value.