u8x8::collect_sum
It occurred to me after writing mask8x8::count_true that of course I'd
really just written a function for summing together all of the elements of
the vector, except that in a mask the maximum value in each element is 1.

So now that same functionality is in u8x8::collect_sum, and the mask8x8
function is just a thin wrapper around that.