u8x8::collect_sum was supposed to be u8x8::reduce_sum
I wrote this "sum the elements of the vector" function right after I'd been
thinking of a nearby problem that involved collecting from an iterator and
so I absent-mindedly gave this function the wrong name.

This should be called "reduce_sum" both because that's consistent with
the likes of core::simd::num::SimdFloat::reduce_sum, and because "reduce"
is just the correct word for this kind of operation.